requirements.txt
# Core CEP Engines
faust-streaming==0.10.4
streamz==0.6.4

# Kafka and messaging
kafka-python==2.0.2
confluent-kafka==2.3.0

# Data processing
pandas==2.1.1
numpy==1.25.2
scikit-learn==1.3.0

# Monitoring and metrics
prometheus-client==0.18.0
psutil==5.9.5

# Statistical analysis
scipy==1.11.3
statsmodels==0.14.0

# Visualization
matplotlib==3.8.0
seaborn==0.12.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Async support
aiohttp==3.8.5
aiokafka==0.8.1

# CLI and utilities
click==8.1.7
tqdm==4.66.1
colorama==0.4.6

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.9.1
flake8==6.1.0
mypy==1.5.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# RocksDB backend for Faust
python-rocksdb==0.7.0

# Serialization
msgpack==1.0.7
